initial commit for project
This commit puts some initial build infrastructure in place, provides a simple
script for installing schema, and a simple test that inserts and retrieves
metric data.

A running instance of Cassandra 2.x is required for installing the schema and
for executing automated tests. Installing the schema is done with,

mvn groovy:execute -Dsource=src/main/script/schema.groovy