add support for configuring the module
The following can now be configured via JSON configuration object that is
passed to the module during initialization,

* nodes
An array of initial cluster contact points. Defaults to 127.0.0.1.

* cluster
The Cassandra cluster name. Defaults to rhq.

* cqlPort
The port for sending client requests to Cassandra. Defaults to 9042.

* httpPort
The port on which to listen for http requests. Defaults to 7474.