some initial multi-tenancy support when pushing data
I have updated the URI for pushing data for a single metric to include the
tenant id. The URI along with the body now looks like,

POST /rhq-metrics/{tenantId}/numeric/{metricId}

[
  {"timestamp": 1416141343793, "value": 2.1},
  {"timestamp": 1416341343793, "value": 2.2},
  {"timestamp": 1416541343793, "value": 2.3}
]

The request body is once again the same as what we currently have in master. If
the tenant does not exist, no data is inserted, and an HTTP 403 status is
returned along with an error message.