add end point for updating numeric meta data
Numeric meta data can be updated as follows,

PUT /{tenantId}/metrics/numeric/{id}/meta

// request body
{
  "attr1": "value 1",
  "attr2: "value 2",
  "[delete]": ["attr3", "attr4"]
}

attr1 and attr2 will be added or replaced if they already exist. [delete] is a
special key which specifies any keys to remove. If you want to delete the meta
data in its entirety, that will be done with a,

DELETE /{tenantId}/metrics/numeric/{id}/meta

Deleting the meta data in its entirety as well as support for availability will
come in subsequent commits.