make the webapp work again with the wildfly maven plugin
The recent change to use Jackson 2 as the JSON provider caused some errors when
deploying with the wildfly maven plugin. I noticed that the plugin deploys an
earlier version of WF. I upgraded the plugin and that resolved the errpr. I
also need to configure the plugin to deploy using the C* backend since a
many of the new core APIs are not implemented for the memory backend.