[HWKMETRICS-8] refactor how metric tags are deleted
Metric tags were previously deleted with,

PUT /{tenantId}/metrics/numeric/{id}/meta

a special key was used to indicate the tags to delete. Deletes are now done at,

DELETE /{tenantId}/metrics/numeric/{id}/meta/t1_name:t1_value$t2_name:t2_value

The tags are encoded in the url. Both the tag name and value need to be
specified. If the tag has no value, then null should be specified. The
name/value pairs are delimited by $ which wil be a reserved character. It will
not be allowed in tag or in metric names.