REST: eagerly load effective configuration
Effective configuration can be loaded on startup: instead of looping on default and config file properties,
we can loop on configuration keys and check if value is present in default, config file, or sys prop.

This eliminates the need for maintaining a set of already read system properties.

Also, simplified config file and property loading code.