[HWKMETRICS-41] refactor response handling
I first updated SimpleDataCallback to handle the various return types which
include Metric, List, and Map. This resulted in some complicated logic in
SimpleDataCallback which involves a lot of type checking. SimplDataCallback is
and was brittle because a REST end point could easily break if a new return
type is added and SimpleDataCallback is not updated accordingly. I opted for a
more functional approach.