HWKMETRICS-67 Influx endpoint: support pretty print query parameter
PrettyFilter checks if the output should be formatted and PrettyInterceptor intercepts JSON response to format it.

I've tried to find how to interact with the RestEasy Jackson mappers directly but the solution would have been hard to maintain.

Here there's a little overhead due to parsing of the entity stream. But this will only be used by humans for a fraction of requests.