New flag to prevent Wildfly to finish the deployment before MetricsService is started
After a minute, we let the deployment finish even if the metrics service is not started yet.
Initialization process will continue in the background, until the deployment is stopped.

rest-tests and ptrans itests setups make use of the new flag to avoid failing the build if
the metrics service takes some time to get ready.