[HWKMETRICS-113] removing jackson and swagger deps from core
It turns out that we were still pulling in jackson transitively from
swagger-core. There were still some jackson annotation in AvailabilityType and
in Tenant that had to be dealt with.

I have removed the swagger annotations from classes in core-api because the
documentation was not generated any way. I will see if I can address that in
subsequent commits, or it may need to be done under a separate ticket.