[HWKMETRICS-130] refactor task execution
Prior to this commit the TaskType class provided a factory for producing an
object to exectue tasks. There are better ways to do this with RxJava. Now
clients register a subscriber with TaskService, and that subscriber is notified
of tasks to execute.

The code for computing rates has been pulled out of MetricsServiceImpl and put
into a new class, GenerateRate. This decouples MetricsServiceInmpl from the
computation of rates which will make it a lot easier to execute these various
components in isolation.