Updated MetricsSender to use the buffer from clients/commons
This verticle consumes metrics published on the bus and inserts them in a buffer.
Metrics are sent as soon as the buffer is larger than the batch size.
If servers are idle then the buffer is flushed, regardless of its size.

When batches fail, the corresponding metrics are re-inserted in the buffer.