[HWKMETRICS-199] initia support for virtual clock in REST API tests
H-Metrics will use the virtual clock when the system property
hawkular.metrics.use-virtual-clock is set to true. It will cause the
VirtualClockHandler to be deployed. It provides endpoints for fetching,
setting, and incrementing the clock.

There is still some work to do to make tests more repeatable. Tests still have
to spin for some short delay to allow time for task execution to complete. This
of course is problematic because there is no way to know how long to block. We
need additional endpoints that provide functionality that are used in other
tests. That is, tests essentially wait for notifications from the task
scheduler that it has finished work for a time slice.