[HWKMETRICS-199] add an endpoint to block until scheduler finishes work
There is a new endpoint to block for a specified duration. On the server side,
we subscribe to the task schedulers time slices observable which emits a
timestamp at the completion of each time slice. That is, a timestamp is emitted
when all of the work for the time slice is finished. A request like,

  GET /hawkulr/metrics/clock/wait?duration=5min

will block the request until the task scheduler has completed the work for five
time slices.