[HWKMETRICS-185] fix jboss-as plugin config and blocking subscriber code
There are three changes with this commit. First, the jboss-as plugin seems to
ignore the <javaOpts> config property. It does recognize <jvmArgs>.

Secondly, there was a subtle JSON mapping issue. When creating a metric with no
tags, MetricDefinition.tags is null and not an empty map. This was leading to a
NPE. I guard against it in MetricDefinition.getTags().

Lastly, in both GaugeHandler.createGaugeMetric and
AvailabilityHandler.createAvailabilityMetric we had,

    observable.toBlocking().last();

The Observable does not emit any items so the last() call results in a
NoSuchElementException. Changing it to lastOrDefault(null) fixes it.