some minor updates to run/test against C* 2.2.0
With this commit, everyone should be using Cassandra 2.2.0 in their local dev
environments.

In additon to having Travis use C* 2.2.0, I had to make some changes around
batch statements. When testing locally with a 2.2.0 cluster, I was getting some
failures about exceeding the batch_size_fail_threshold_in_kb setting defined
in cassandra.yaml. It defaults to 50 KB. BatchStatementTransformer had a check
on the number of statements in a batch. This is not really useful though
because the check needs to be on the overall size, not the number of statements
in the batch. I am going to create a ticket to address our usage of batch
statements. The conventional wisdom in general is to avoid them, and it makes
sense for adhere to that.