[HWKMETRICS-292] initial support for aggregating data across multiple metrics
This commit also includes a change to NumericDataPointCollector to facilitate
testing. The order of input of values will effect the results of percentiles.
This makes testing the new API difficult because we are fetching data points
from multiple metrics concurrently which means the order of input is
non-deterministic. The test hook allows for using a different implementation in
which the order of inputs does not matter.