HWKMETRICS-315 Bus integration: events randomly or never posted
When Metrics was started after the Pinger, the Pinger would accumulate data to post before sending it all at once eventually.
Because there were not enough threads to handle the peak load, RxJava tried to apply backpressure but PublishSubject does not support it by nature.

The fix consists in:
- buffering inserted metric to avoid submitting too much to the io scheduler
- adding a buffer-on-backpressure strategy

Note that the problem was emphasized before as we used to schedule sending on a wrapped ManagedExecutorService. Now we use the #io scheduler (change already in master).

I've successfully tested the changes on Hawkular server.