[HWKMETRICS-337] initial commit for Bus class and new arquillian integration tests
Other than moving the bus integration code to the new hawkular-metrics-bus
module, the code is largely unchanged. I moved the integration code to the new
bus module to clearly separate the purpose of this code which also makes it
easier to test in isolation. There are a couple arquillian integration tests
that basically test the same functionality as the out-of-container tests
currently in hawkular-component.

hawular-component no longer contains any java source code and just builds the
WAR overlay. In a follow up commit, I will replace the tests in this module
with full, end-to-end tests that include alerts integration. I thihk it makes
sense to include alerts integration in the tests since alerts is the primary
consumer of the data we are publishing onto the bus.

There was code in metrics-api-common that metrics-bus depends on, like the
@ServiceReady annotation. I moved those classes into another new module,
hawkular-metrics-api-util, which is probably not the best name considering the
reasons for the change. metrics-api-common depends on metrics-core-service. I
wanted to avoid that dependency in metrics-bus, partly to facilitate testing
and also to help keep things loosely coupled.

The api-util module also provides a place to share common code needed the core
REST API and any integration components. The ServiceReadyEvent class is one
such example. MetricsService was previously the event object sent when
ServiceReady was fired. We only need the observable that emits inserted metrics
so I introduced ServiceReadyEvent which encapsulates that observable and could
include anything else that we later want to pass in the event.