fix hawkular runtime bus dependency
My changes for HWKMETRICS-337 wound up putting the hawkular bus JAR in the
hawkular-component WAR while still also declaring the dependency on the nest.
This commit takes it out of the WAR and declares the dependency on the bus
module that is packaged with the server. We no longer declare a depdendency on
the nest module. That was really only for consuming the bus.