[HWKMETRICS-361] add bootstrap script to properly initialize schema
The previous code in cassalog.groovy that had some conditional logic based on
the version of Hawkular Metrics had some holes in the logic. bootstrap.groovy
has been added to handle new installs and upgrades. Any schema changes can
then be placed in cassalog.groovy without having to worry about whether it is
a new install or an upgrade.