[BZ 1316275] add support for altering gc_grace_seconds in data table
A new config property has been added. It is called DATA_GC_GRACE_SECONDS. At
start up we query the system tables to get the current value of
gc_grace_seconds for the data table. If the current value does not match the
config value, which defaults to 1 day, we update gc_grace_seconds.