HWKMETRICS-385 Deleting tags should only require the tag keys and not the values
Added a new JAX-RS param class. It allows to map a list of plain names mixed with name:value pairs (for backwared compatibility)
Updated the core service implementation so that tag values are loaded and from metric index. This is needed otherwise we can't delete rows from the tags index table (tvalue is part of the primary key).