HWKMETRICS-390 Endpoint for fetching rate data points need to support standard query params and sort behavior
I added a method in Order.java which factors out the logic to determine a default order. I updated the considering endpoints.

I change the get rate endpoint to support the new parameters. While its bucket parameters are deprecated and not removed, we need to check that they can't be set if order and limit are.

While updating the code in MetricsServiceImpl to adapt to the new  findRate signature, I extracted the code of #bucketize methods and created separated Transformer classes. Using these transformers with #compose is a more natural RxJava construct.