[HWKMETRICS-360] initial commit for new job-scheduler module
This commit includes some initial functionality for scheduling jobs. I have
also pulled DateTimeService out into a separate module so that it can be used
by both metrics-core-service and by job-scheduler.

[HWKMETRICS-360] fix merge conflicts from rebase and use config service

I have also updated ConfigurationService to use LOCAL_QUORUM consistency and
added a method for updating a single property.

[HWKMETRICS-360] initial support for executing single execution jobs

This commit adds support for executing single execution or non-repeating jobs.
The job scheduler looks for and executes jobs by minute time slices. Every
minute a new task is kicked off to find jobs to execute. The active, i.e.,
oldest time slice is stored in the sys_config table. The task will first check
for jobs in that time slice. If there are available jobs, the scheduler will
obtain the necessary locks and execute the jobs. If there are no jobs or if all
of the jobs are already running the scheduler will look for jobs in following
time slices. This is still pretty rough around the edges.

[HWKMETRICS-360] initial support for executing repeating jobs

There are a few important changes for handling repeating jobs. The first and
probably most obvious is having to reschedule the job. This involves updating
the jobs and the jobs_status tables.

Next, when job execution is done, we need to release the job lock. We use
thread names for lock values. Since we don't know which thread is running when
a lock is acquired, we need to remember the lock value in some fashion. The
current solution is not so great as it basically involves a global variable.

Lastly, we need to remove the job id from the activeJobs set created in the
start method. If we do not do this, we could block subsequent executions of the
job.

Some minor changes were made to LockManager. There are methods acquiring locks
and for acquiring permits. Permits are used for shared or multi access as in
the case of scheduling jobs. We want to allow multiple clients schedule jobs
concurrently. For executing a job, we want an exclusive lock that has a single
owner. Right now I am using the same table for both of these. We'll see how it
goes.

[HWKMETRICS-360] try to make tests more robust

Previously tests were using a shared job scheduler instance. I started running
into problems once I added tests for repeating jobs. Jobs set up from one test
method were running in later tests and causing problems. I decided that a new
job scheduler instance with a clean db per test would be better.

[HWKMETRICS-360] finally work around deadlock issues

I still do not understand but in my unit tests I was hitting deadlock issues.
I had to move the execution of most queries off of the computation scheduler
and onto a different scheduler.

[HWKMETRICS-360] make the job factories return a Completable

Completable is a better type to use for the job execution code. It is a more
natural fit for async execution of code that involves side effects.

committing local edits for others to see

[HKKMETRICS-360] some minor clean up, remove unneeded code

[HWKMETRICS-360] fix checkstyle, license errors

[HWKMETRICS-360] fix version numbers

[HWKMETRICS-360] turn on debug logging for travis, configure RejectedExecutionHandler

[HWKMETRICS-360] initialze TestScheduler for each test

[HWKMETRICS-360] fix timing issue in executeLongRunningSingleExecutionJob()

[HWKMETRICS-360] fix missing dependency after rebase

[HWKMETRICS-360] add missing import

[HWKMETRICS-360] more refactoring/debugging

[HWKMETRICS-360] get post job execution clean up working with Completable

more debugging

increase timeout to avoid race condition

clean up obsolete code and add some comments

This is squash commit of several small changes in an effort to eliminate
intermittent test failures.

fix version for datetime service

remove duplicate dependency declaration

serialize access to activeTimeSlices cache

remove unused debug code that was accidentally committed

remove legacy, unused task scheduler code

remove configuration service since it is no longer used in scheduler

initial commit for JobsService

JobsService lives in metrics-core-service. It serves a couple purposes. It
encapsulates the scheduler and controls its lifecycle. JobsService will also
provide APIs for submitting and canceling jobs.

fix license headers

[HWKMETRICS-360] add some debug logging and use concat instead of merge

SchedulerImpl.doJobExecution() was previously using Completable.merge which
allows for out of order execution of the Completable arguments. In testing I
discovered that notifications of job completion were being sent before job
execution completed. This was due to using merge. Completable.concat fixes
this.

a little clean up, trying to get github to see my previous commit

[HWKMETRICS-360] use Observable version of updateActiveTimeSlices to avoid eager execution