[HWKMETRICS-360] initial support for executing single execution jobs
This commit adds support for executing single execution or non-repeating jobs.
The job scheduler looks for and executes jobs by minute time slices. Every
minute a new task is kicked off to find jobs to execute. The active, i.e.,
oldest time slice is stored in the sys_config table. The task will first check
for jobs in that time slice. If there are available jobs, the scheduler will
obtain the necessary locks and execute the jobs. If there are no jobs or if all
of the jobs are already running the scheduler will look for jobs in following
time slices. This is still pretty rough around the edges.