[HWKMETRICS-360] initial support for executing repeating jobs
There are a few important changes for handling repeating jobs. The first and
probably most obvious is having to reschedule the job. This involves updating
the jobs and the jobs_status tables.

Next, when job execution is done, we need to release the job lock. We use
thread names for lock values. Since we don't know which thread is running when
a lock is acquired, we need to remember the lock value in some fashion. The
current solution is not so great as it basically involves a global variable.

Lastly, we need to remove the job id from the activeJobs set created in the
start method. If we do not do this, we could block subsequent executions of the
job.

Some minor changes were made to LockManager. There are methods acquiring locks
and for acquiring permits. Permits are used for shared or multi access as in
the case of scheduling jobs. We want to allow multiple clients schedule jobs
concurrently. For executing a job, we want an exclusive lock that has a single
owner. Right now I am using the same table for both of these. We'll see how it
goes.