[HWKMETRICS-393] add endpoint for fetching raw data from multiple gauges
This new endpoint goes against REST design principles by using POST for a
query. This has been done in order to avoid problems with complex and long
query strings in the URL.

This endpoint is also a bit different in that it streams the resposne. The
reason for this is that the client can request an arbitrary amount of data
from an arbitrary number of metrics.