[HWKMERTICS-393] several changes based on PR review
I eliminated most of the duplicate code by pushing it up into a new parent
class, MetricsServiceHandler. The name sucks, but both BaseHandler and
MetricHandler are already taken.

I have added a bunch of swagger annotations for the new endpoints.

Lastly, I put the CountDownLatch.countDown() call in a finally block to ensure
it happens to prevent blocking indefinitely.