HWKMETRICS-429 Let ptrans connect to Hawkular-Services and Openshift-Metrics
Since Vert.x Hawkular Metrics already knows how to connect to the backend, there is no point in maintaining the same kind of functionality in ptrans.
So now the protocol servers send json to the event bus, and the Vert.x Metrics SPI impl will take care of forwarding the datapoints.

Note that since Vert.x 3.3.0, the SPI impl knows how to connect to standalone Metrics, old Hawkular servers, Hawkular Services and Openshift-Metrics.
So we only have to configure it properly.

Also, use Maven Shade Plugin to create the fat JAR

Assembly plugin is not recommended for this usage (because, among other
things, you can't control the output in case of conflicts)