[HWKMETRICS-446] Add REST integration tests for tenant deletion
There was a bit of work involved because we need test hooks for the job
scheduler in our REST tests just like we do in the core java integration
tests. There is a new module, metrics-api-jaxrs-test, which is a WAR overlay
and deploy and additional resource/handler named JobSchedulerHandler. With
this commit it provides endpoints for changing the test clock.