[HWKMETRICS-1] first pass at using ISPN to do roll up computations
The roll up job is now configured to run every minute. When data points are
persisted, we also write them to a cache, which is configured to be
distributed. The roll up job fetches the data from the cache, aggregates it,
and persisted the roll up (i.e., bucketed) data point. After the data is
written to Cassandra, it is removed from the cache.