[HWKMETRICS-1] compute 1 minute and 5 minute rollups
For the 1 minute rollups, we maintain a cache of the raw data points and
aggregate them every minute. For other, lower resolution rollups such as 5min,
we maintain a cache of NumericDataPointCollectors. The collectors get updated
every time we aggregate the raw data. When the 5 minute time slice is up, we
obtain a bucket point from each collector and persist each of them.