[HWKMETRICS-465] New POST endpoints gauges/stats/query and counters/stats/query
This is part of several commits to harmonize the REST API when querying on multiple metrics.
With those new endpoints it is now possible to query on x/stats/query, in a similar way than it was already possible on x/raw/query and x/rate/query.

Also added integration tests, fix missing 'tags' in hashcode