[HWKMETRICS-505] Several changes to improve error handling
This commit includes a bunch of changes.

* Store job details in scheduled_jobs_idx
I previously only stored the job id in scheduled_jobs_idx and fetched the
details from the jobs table. This eliminates the extra read. I also am no
longer storing the job details in the jobs table. Right now it doesn't really
serve a purpose, and it has actually made things a little more complex because
there are times when the next trigger gets persisted and the job for the
current trigger still has to be retried.

* Do not updated finished_jobs_idx until other post-execution steps are done
I previously updated finished_jobs_idx right after the job was done executing.
If a subsequent post-execution step failed, namely reschulding, the job would
never run again because the computeRemainingJobs() method which determines
eligible jobs to execute filters out jobs that are already in finished_jobs_idx
for the time slice in question.

* Add status column to scheduled_jobs_idx
Rather than updating finished_jobs_idx right away, I now update this new status
column. If the job is then retried after a failure we first check to see if
the status flag is set. If it is we skip job execution and move right to the
post-execution steps.

* Add error logging
I have added a bunch of doOnError(Throwable) calls in the post-execution steps
for improved debugging to make it easier to see what exactly fails.

* Make sure active jobs cache is updated
Previously if there was an error during post-execution, the job id would not
get removed from the active jobs cache. This would prevent subsequent execution
of the job since we filter out those jobs in the active jobs cache when
determining the eligible jobs. A server restart would have been required. Now
I make sure the cache gets updated in a doOnError(Throwable) callback that is
at the end of the post-execution call chain.

[HWKMETRICS-505] fix compiler error]

Forgot to remove reference to uncommitted, PoC code.

add license header

[HWKMETRICS-505] fix broken tests

Tests broke since we are no longer updating the jobs table.

fix compile error

[HWKMETRICS-505] changes based on PR review

license header